requirements.txt
# Core dependencies
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.24.0

# NLP and ML packages
spacy>=3.7.2
nltk>=3.8.1
textblob>=0.17.1
vaderSentiment>=3.3.2
transformers>=4.35.0
torch>=2.1.0
scikit-learn>=1.3.0

# Advanced NLP models
keybert>=0.7.0
bertopic>=0.15.0
sentence-transformers>=2.2.2

# Text analysis utilities
textstat>=0.7.3
emot>=3.1